chores/file-rename: Rename controller model service file naming structure
