+++ src/controllers/movies.controller.js
@@ -1,6 +1,6 @@
 import HttpStatus from 'http-status-codes';
 
-import * as movieService from '@services/movieService';
+import * as movieService from '@/services/movie.service';
 
 import { DEFAULT_PAGE_NUMBER, DEFAULT_PAGE_SIZE } from '@/constants/constant';
