+++ src/services/rent.service.js
@@ -1,6 +1,6 @@
 import Boom from '@hapi/boom';
 
-import Rent from '@models/rent';
+import Rent from '@/models/rent.model';
 
 /**
  * Create new rent.
