+++ src/services/role.service.js
@@ -1,6 +1,6 @@
 import Boom from '@hapi/boom';
 
-import Role from '@models/role';
+import Role from '@/models/role.model';
 
 /**
  * Get all roles.
